# src/main/java/io/rosecloud/iam/api/OperatorSetupCompleteRequest.java
package io.rosecloud.iam.api;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Pattern;

public record OperatorSetupCompleteRequest(@NotBlank String setupToken, String totpCode) {}